+++ addons/main/config.cpp
@@ -77,6 +77,7 @@ class CfgFunctions
             class loadoutMenu {};
             class loadoutWatch {};
             class loadoutSample {};
+            class squadLeaderWatch {};
             class squadLeaderSample {};
             class placeOf {};
             class placePosition {};